Crate mariadb_sys

Bindings module

Autogenerated bindings to C interfaces for Rust

Re-exports

pub use self::enum_repertoire_t::Type as my_repertoire_t;

pub use self::enum_collation_id_type::Type as my_collation_id_type_t;

pub use self::enum_collation_name_mode::Type as my_collation_name_mode_t;

pub use self::key_cache_type::Type as KEY_CACHE_TYPE;

pub use self::enum_table_category::Type as TABLE_CATEGORY;

Modules

Bitmap_Iterator__bindgen_ty_1

Compare_keys

Derivation

Item_result

Lex_exact_charset_extended_collation_attrs_st_Type

Lex_extended_charset_extended_collation_attrs_st_charset_type_t

Lex_extended_collation_st_Type

Log_event_data_type__bindgen_ty_1

MDL_key_enum_mdl_namespace

MDL_ticket_enum_psi_status

MDL_wait_enum_wait_status

MDL_wait_for_subgraph_enum_deadlock_weight

PSI_table_io_operation

SEQUENCE_seq_init

Sql_used__bindgen_ty_1

TABLE_LIST_enum_open_strategy

TABLE_LIST_prelocking_types

TABLE_SHARE_enum_v_keys

Table_map_iterator__bindgen_ty_1

Table_type

cache_type

chf_create_flags

column_definition_type_t

date_mode_t_value_t

enum_alter_inplace_result

enum_binlog_command

enum_binlog_func

enum_collation_id_type

enum_collation_name_mode

enum_dynamic_column_type

enum_field_types

enum_ft_token_type

enum_ftparser_mode

enum_mdl_duration

enum_mdl_type

enum_mysql_show_type

enum_mysql_stmt_state

enum_mysql_timestamp_type

enum_nullability

enum_open_type

enum_plugin_load_option

enum_query_type

enum_repertoire_t

enum_schema_table_state

enum_schema_tables

enum_show_open_table

enum_stats_auto_recalc

enum_table_category

enum_table_ref_type

enum_vcol_update_mode

file_opt_type

flush_type

ha_choice

ha_extra_function

ha_key_alg

ha_option_type

ha_panic_function

ha_rkey_function

ha_stat_type

ha_storage_media

handler_batch_mode_t

handler_init_stat

interval_type

key_cache_type

legacy_db_type

loglevel

mysql_option

mysql_status

open_frm_error

partition_value_print_mode_t

privilege_t

protocol_send_type_t

row_type

st_order_enum_order

thd_kill_levels

thr_lock_type

tmp_table_type

vers_kind_t

vers_system_time_t

Structs

ACL_internal_schema_access

ACL_internal_table_access

ALL_READ_COST

AUTHID

Alter_info

Alter_inplace_info

Class describing changes to be done by ALTER TABLE. Instance of this class is passed to storage engine in order to determine if this ALTER TABLE can be done using in-place algorithm. It is also used for executing the ALTER TABLE using in-place algorithm.

Alter_inplace_info_Rename_key_pair

Old and new index names. Used for index rename.

Arg_comparator

Binary_string

Storage for strings with both length and allocated length. Automatically grows on demand.

Bit_addr

Bitmap_Iterator

Blob_mem_storage

Class is used as a BLOB field value storage for intermediate GROUP_CONCAT results. Used only for GROUP_CONCAT with DISTINCT or ORDER BY options.

COND_EQUAL

Cache_flip_event_log

Charset

Charset_collation_context

Charset_collation_map_st

Charset_collation_map_st_Elem

Charset_collation_map_st_Elem_st

Column_definition

Column_definition_attributes

Column_derived_attributes

Compare_identifiers

Compare_table_names

Conv_source

Cost_estimate

Create_field

DTCollation

“Declared Type Collation” A combination of collation and its derivation.

Default_object_creation_ctx

Default_object_creation_ctx – default implementation of Object_creation_ctx.

Discrete_interval

Dynamic_array

ErrBuff

Event_log

Exec_time_tracker

Field

Field_translator

File_option

File_parser

GRANT_TABLE

Gap_time_tracker

HA_CREATE_INFO

This struct is passed to handler table routines, e.g. ha_create(). It does not include the “OR REPLACE” and “IF NOT EXISTS” parts, as these parts are handled on the SQL level and are not needed on the handler level.

Handler_share

Base class to be used by handlers different shares

Handler_share__bindgen_vtable

IO_AND_CPU_COST

IS_table_read_plan

I_P_List

Intrusive parameterized list.

I_P_List_counter

Element counting policy class for I_P_List which provides basic element counting.

I_P_List_iterator

Iterator for I_P_List.

I_P_List_no_push_back

A null insertion policy class for I_P_List to be used in cases when push_back() operation is not necessary.

I_P_List_null_counter

Element counting policy class for I_P_List to be used in cases when no element counting should be done.

Index_hint

Index_statistics

Item

Item_bool_func2

Item_bool_rowready_func2

Item_cache

Item_char_typecast

Item_copy

Item_date_typecast

Item_datetime_typecast

Item_decimal_typecast

Item_double_typecast

Item_field

Item_float_typecast

Item_func_abs

Item_func_between

Item_func_div

Item_func_hash

Item_func_hex

Item_func_hybrid_field_type

Item_func_in

Item_func_int_val

Item_func_min_max

Item_func_minus

Item_func_mod

Item_func_mul

Item_func_neg

Item_func_or_sum

Item_func_plus

Item_func_round

Item_func_signed

Item_func_unsigned

Item_in_subselect

Item_param

Item_subselect

Item_sum_avg

Item_sum_hybrid

Item_sum_sum

Item_sum_variance

Item_time_typecast

KEY_FIELD

KEY_PAIR

Structure describing changes to an index to be caused by ALTER TABLE.

Key_part_spec

LEX

last_leaf_for_name_resolutioning support.

LEX_USER

LF_DYNARRAY

LF_PINBOX

LF_PINS

Lex_context_collation

Lex_cstring

Lex_cstring_with_compare

Lex_exact_charset

Lex_exact_charset_extended_collation_attrs_st

Lex_exact_charset_opt_extended_collate

Lex_exact_collation

Lex_extended_charset_extended_collation_attrs_st

Lex_extended_collation

Lex_extended_collation_st

Lex_field_type_st

Lex_length_and_dec_st

Lex_opt_context_charset_st

List

List_Iterator

Log_event_data_type

Longlong

Longlong_hybrid

MDL_context

Context of the owner of metadata locks. I.e. each server connection has such a context.

MDL_context_owner

An interface to separate the MDL module from the THD, and the rest of the server code.

MDL_context_owner__bindgen_vtable

MDL_key

@sa MDL_key.

MDL_lock

MDL_request

A pending metadata lock request.

MDL_savepoint

Savepoint for MDL context.

MDL_ticket

A granted metadata lock.

MDL_wait

A reliable way to wait on an MDL lock.

MDL_wait_for_graph_visitor

An abstract class for inspection of a connected subgraph of the wait-for graph.

MDL_wait_for_graph_visitor__bindgen_vtable

MDL_wait_for_subgraph

Abstract class representing an edge in the waiters graph to be traversed by deadlock detection algorithm.

MDL_wait_for_subgraph__bindgen_vtable

MYSQL_LOG

MY_STRCOPY_STATUS

Name

Name_resolution_context

Native

Natural_join_column

OPTIMIZER_COSTS

Object_creation_ctx

Object_creation_ctx – interface for creation context of database objects (views, stored routines, events, triggers). Creation context – is a set of attributes, that should be fixed at the creation time and then be used each time the object is parsed or executed.

Object_creation_ctx__bindgen_vtable

Open_table_context

PARTITION_STATS

PSI_cond

Interface for an instrumented condition. This is an opaque structure.

PSI_file

Interface for an instrumented file handle. This is an opaque structure.

PSI_metadata_lock

Interface for an instrumented metadata lock. This is an opaque structure.

PSI_mutex

Interface for an instrumented mutex. This is an opaque structure.

PSI_rwlock

Interface for an instrumented rwlock. This is an opaque structure.

PSI_socket

Interface for an instrumented socket descriptor. This is an opaque structure.

PSI_stage_info_v1

Stage instrument information. @since PSI_VERSION_1 This structure is used to register an instrumented stage.

PSI_table

Interface for an instrumented table handle. This is an opaque structure.

PSI_table_locker

Interface for an instrumented table operation. This is an opaque structure.

PSI_table_locker_state

State data storage for @c start_table_io_wait_v1_t, @c start_table_lock_wait_v1_t. This structure provide temporary storage to a table locker. The content of this structure is considered opaque, the fields are only hints of what an implementation of the psi interface can use. This memory is provided by the instrumented code for performance reasons. @sa start_table_io_wait_v1_t @sa start_table_lock_wait_v1_t

PSI_table_share

Interface for an instrumented table share. This is an opaque structure.

PSI_thread

Interface for an instrumented thread. This is an opaque structure.

Protocol

Pushdown_derived

Query

The structure describing various parts of the query

Query_arena

Range_rowid_filter_cost_info

Record_addr

Row_definition_list

Rowid_filter

SEQUENCE

SEQUENCE is in charge of managing the sequence values. It’s also responsible to generate new values and updating the sequence table (engine=SQL_SEQUENCE) trough it’s specialized handler interface.

SJ_MATERIALIZATION_INFO

SORT_FIELD_ATTR

ST_FIELD_INFO

Schema

Schema_specification_st

A helper struct for schema DDL statements: CREATE SCHEMA [IF NOT EXISTS] name [ schema_specification… ] ALTER SCHEMA name [ schema_specification… ]

Security_context

Select_limit_counters

Show_Type

SplM_opt_info

Sql_alloc

Sql_used

String

TABLE

TABLE_LIST

TABLE_OPT_RANGE

TABLE_SHARE

This structure is shared between different table objects. There is one instance of table share per one table in the database.

TABLE_SHARE_period_info_t

System versioning and application-time periods support.

TABLE_STATISTICS_CB

TDC_element

THD

TMP_TABLE_PARAM

Table_function_json_table

Table_map_iterator

Table_path_buffer

Table_period_info

Table_period_info_start_end_t

Table_scope_and_contents_source_pod_st

A helper struct for table DDL statements, e.g.: CREATE [OR REPLACE] [TEMPORARY] TABLE [IF NOT EXISTS] tbl_name table_contents_source;

Table_scope_and_contents_source_st

Table_statistics

Table_triggers_list

Temporal_Status

Temporal_Warn

Type_aggregator

Type_aggregator_Pair

Type_all_attributes

Type_all_attributes__bindgen_vtable

Type_cmp_attributes

Type_cmp_attributes__bindgen_vtable

Type_collection

Type_collection__bindgen_vtable

Type_handler

Type_handler__bindgen_vtable

Type_handler_hybrid_field_type

A handler for hybrid type functions, e.g. COALESCE(), IF(), IFNULL(), NULLIF(), CASE, numeric operators, UNIX_TIMESTAMP(), TIME_TO_SEC().

Type_numeric_attributes

Type_std_attributes

A class to store type attributes for the standard data types. Does not include attributes for the extended data types such as ENUM, SET, GEOMETRY.

Type_std_attributes_Single_coll_err

Typelib

USER_AUTH

Unknown_key_hook

This hook used to catch no longer supported keys and process them for backward compatibility.

Unknown_key_hook__bindgen_vtable

Vers_history_point

Vers_parse_info

View_creation_ctx

View_creation_ctx – creation context of view objects.

Virtual_column_info

Wait_for_flush

Class representing the fact that some thread waits for table share to be flushed. Is used to represent information about such waits in MDL deadlock detector.

With_element

_IO_FILE

_IO_codecvt

_IO_marker

_IO_wide_data

__BindgenBitfieldUnit

__BindgenUnionField

__atomic_wide_counter

__atomic_wide_counter__bindgen_ty_1

__pthread_cond_s

__pthread_internal_list

__pthread_mutex_s

__pthread_rwlock_arch_t

_db_code_state_

_db_stack_frame_

_ft_vft

base_list

binlog_cache_data

casefold_info_char_t

casefold_info_st

charset_info_st

cmp_item

date_mode_t

derived_handler

embedded_query_result

engine_option_value

fileinfo

group_by_handler

group_by_handler__bindgen_vtable

ha_copy_info

ha_field_option_struct

ha_handler_stats

ha_index_option_struct

ha_statistics

ha_table_option_struct

handler

The handler class is the interface for dynamically loadable storage engines. Do not add ifdefs and take care when adding or changing virtual functions to avoid vtable confusion

handler__bindgen_vtable

handler_bridge

Wrapper that can expose a C vtable as a C++ class

handler_bridge_vt

A C representation of the handler class.

handlerton

handlerton_discovered_list

handlerton_discovered_list__bindgen_vtable

ilist_node

in_vector

inplace_alter_handler_ctx

In-place alter handler context.

inplace_alter_handler_ctx__bindgen_vtable

list_node

list_node - a node of a single-linked list. @note We never call a destructor for instances of this class.

my_charset_handler_st

my_charset_loader_st

my_collation_handler_st

my_contraction_list_t

my_contraction_t

my_decimal

my_decimal class limits ‘decimal_t’ type to what we need in MySQL.

my_match_t

my_uca_2bytes_item_t

my_uca_contraction_hash_t

my_uca_level_booster_t

my_uca_level_info_st

my_uca_weight2_t

my_uni_idx_st

online_alter_cache_data

partition_info

pthread_cond_t

pthread_mutex_t

pthread_rwlock_t

select_handler

select_unit

sequence_definition

sequence_definition is used when defining a sequence as part of create

sql_service_st

st_bitmap

st_cond_statistic

st_decimal_t

intg is the number of decimal digits (NOT number of decimal_digit_t’s !) before the point frac is the number of decimal digits after the point len is the length of buf (length of allocated space) in decimal_digit_t’s, not in bytes sign false means positive, true means negative buf is an array of decimal_digit_t’s

st_dynamic_array

st_ft_info

st_grant_info

The current state of the privilege checking process for the current

st_grant_internal_info

State information for internal tables grants. This structure is part of the TABLE_LIST, and is updated during the ACL check process. @sa GRANT_INFO

st_ha_check_opt

st_ha_create_table_option

st_handler_buffer

st_hash

st_io_cache

st_io_cache_share

st_join_table

st_key

st_key__bindgen_ty_1

st_key_cache

st_key_cache_funcs

st_key_cache_statistics

st_key_multi_range

st_key_part_info

st_key_range

st_list

st_lock_list

st_maria_plugin

st_mariadb_encryption

Encryption plugin descriptor

st_mem_root

st_my_dir

st_mysql

st_mysql_bind

st_mysql_cond

An instrumented cond structure. @sa mysql_cond_t

st_mysql_const_lex_string

st_mysql_const_unsigned_lex_string

st_mysql_data

st_mysql_field

st_mysql_file

An instrumented FILE structure. @sa MYSQL_FILE

st_mysql_ftparser_boolean_info

st_mysql_ftparser_param

st_mysql_lex_string

st_mysql_methods

st_mysql_mutex

An instrumented mutex structure. @sa mysql_mutex_t

st_mysql_options

st_mysql_options_extention

@todo remove the “extension”, move st_mysql_options completely out of mysql.h

st_mysql_parameters

st_mysql_plugin

st_mysql_prlock

An instrumented prlock structure. @sa mysql_prlock_t

st_mysql_res

st_mysql_rows

st_mysql_rwlock

An instrumented rwlock structure. @sa mysql_rwlock_t

st_mysql_show_var

st_mysql_socket

An instrumented socket.

st_mysql_stmt

st_mysql_stmt_extension

st_mysql_storage_engine

st_mysql_sys_var

st_mysql_time

st_mysql_time_status

st_mysql_value

st_mysql_xid

struct st_mysql_xid is binary compatible with the XID structure as in the X/Open CAE Specification, Distributed Transaction Processing: The XA Specification, X/Open Company Ltd., 1991. http://www.opengroup.org/bookstore/catalog/c193.htm

st_nested_join

st_net

st_order

st_page_range

st_plugin_dl

st_plugin_int

st_ptr_backup

st_range_seq_if

st_reginfo

st_rw_pr_lock_t

Portable implementation of special type of read-write locks.

st_schema_table

st_select_lex

st_select_lex_unit

st_table_field_def

st_table_field_type

st_thr_lock

st_thr_lock_data

st_thr_lock_info

st_typelib

st_used_mem

st_value

st_vio

stat

std_forward_iterator_tag

std_input_iterator_tag

sys_var

sysvar_bool_t

sysvar_common_t

sysvar_double_t

sysvar_enum_t

sysvar_int_t

sysvar_long_t

sysvar_longlong_t

sysvar_set_t

sysvar_str_t

sysvar_uint_t

sysvar_ulong_t

sysvar_ulonglong_t

timespec

uca_info_st

vers_history_point_t

vers_select_conds_t

xid_t

struct xid_t is binary compatible with the XID structure as in the X/Open CAE Specification, Distributed Transaction Processing: The XA Specification, X/Open Company Ltd., 1991. http://www.opengroup.org/bookstore/catalog/c193.htm

Constants

Bitmap_BITS_PER_ELEMENT

CLIENT_PLUGIN_AUTH

CLIENT_PLUGIN_AUTH_LENENC_CLIENT_DATA

ENCRYPTION_FLAG_DECRYPT

ENCRYPTION_FLAG_ENCRYPT

ENCRYPTION_FLAG_NOPAD

ENCRYPTION_KEY_BUFFER_TOO_SMALL

ENCRYPTION_KEY_NOT_ENCRYPTED

ENCRYPTION_KEY_SYSTEM_DATA

ENCRYPTION_KEY_TEMPORARY_DATA

ENCRYPTION_KEY_VERSION_INVALID

ENCRYPTION_SCHEME_BLOCK_LENGTH

ENCRYPTION_SCHEME_KEY_INVALID

HA_ADMIN_ALREADY_DONE

HA_ADMIN_COMMIT_ERROR

HA_ADMIN_CORRUPT

HA_ADMIN_FAILED

HA_ADMIN_INTERNAL_ERROR

HA_ADMIN_INVALID

HA_ADMIN_NEEDS_ALTER

HA_ADMIN_NEEDS_CHECK

HA_ADMIN_NEEDS_UPGRADE

HA_ADMIN_NOT_BASE_TABLE

HA_ADMIN_NOT_IMPLEMENTED

HA_ADMIN_OK

HA_ADMIN_REJECT

HA_ADMIN_TRY_ALTER

HA_ADMIN_WRONG_CHECKSUM

HA_ANY_INDEX_MAY_BE_UNIQUE

HA_AUTO_KEY

HA_AUTO_PART_KEY

HA_BINARY_PACK_KEY

HA_BINLOG_FLAGS

HA_BINLOG_ROW_CAPABLE

HA_BINLOG_STMT_CAPABLE

HA_BIT_PART

HA_BLOB_PART

HA_CACHE_TBL_ASKTRANSACT

HA_CACHE_TBL_NOCACHE

HA_CACHE_TBL_NONTRANSACT

HA_CACHE_TBL_TRANSACT

HA_CAN_BIT_FIELD

HA_CAN_DIRECT_UPDATE_AND_DELETE

HA_CAN_EXPORT

HA_CAN_FORCE_BULK_DELETE

HA_CAN_FORCE_BULK_UPDATE

HA_CAN_FULLTEXT

HA_CAN_FULLTEXT_EXT

HA_CAN_GEOMETRY

HA_CAN_HASH_KEYS

HA_CAN_INDEX_BLOBS

HA_CAN_INSERT_DELAYED

HA_CAN_MEMCMP

HA_CAN_MULTISTEP_MERGE

HA_CAN_ONLINE_BACKUPS

HA_CAN_REPAIR

HA_CAN_RTREEKEYS

HA_CAN_SKIP_LOCKED

HA_CAN_SQL_HANDLER

HA_CAN_TABLES_WITHOUT_ROLLBACK

HA_CAN_TABLE_CONDITION_PUSHDOWN

HA_CAN_VIRTUAL_COLUMNS

HA_CHECK_ALL

HA_CHECK_DUP

HA_CHECK_DUP_KEY

HA_CHECK_DUP_UNIQUE

HA_CHECK_FK_ERROR

HA_CLUSTERED_INDEX

HA_CMP_REF_IS_EXPENSIVE

HA_CONCURRENT_OPTIMIZE

HA_CRASH_SAFE

HA_CREATE_CHECKSUM

HA_CREATE_DELAY_KEY_WRITE

HA_CREATE_INTERNAL_TABLE

HA_CREATE_KEEP_FILES

HA_CREATE_NO_ROLLBACK

HA_CREATE_PAGE_CHECKSUM

HA_CREATE_PRINT_ALL_OPTIONS

HA_CREATE_RELIES_ON_SQL_LAYER

HA_CREATE_TMP_ALTER

HA_CREATE_TMP_TABLE

HA_CREATE_UNIQUE_INDEX_BY_SORT

HA_CREATE_USED_AUTO

HA_CREATE_USED_AVG_ROW_LENGTH

HA_CREATE_USED_CHARSET

HA_CREATE_USED_CHECKSUM

HA_CREATE_USED_COMMENT

HA_CREATE_USED_CONNECTION

HA_CREATE_USED_DATADIR

HA_CREATE_USED_DEFAULT_CHARSET

HA_CREATE_USED_DELAY_KEY_WRITE

HA_CREATE_USED_ENGINE

HA_CREATE_USED_INDEXDIR

HA_CREATE_USED_INSERT_METHOD

HA_CREATE_USED_KEY_BLOCK_SIZE

HA_CREATE_USED_MAX_ROWS

HA_CREATE_USED_MIN_ROWS

HA_CREATE_USED_PACK_KEYS

HA_CREATE_USED_PAGE_CHECKSUM

HA_CREATE_USED_PASSWORD

HA_CREATE_USED_RAID

HA_CREATE_USED_ROW_FORMAT

HA_CREATE_USED_SEQUENCE

HA_CREATE_USED_STATS_AUTO_RECALC

HA_CREATE_USED_STATS_PERSISTENT

HA_CREATE_USED_STATS_SAMPLE_PAGES

HA_CREATE_USED_TRANSACTIONAL

HA_CREATE_USED_UNION

HA_DONT_TOUCH_DATA

HA_DO_INDEX_COND_PUSHDOWN

HA_DO_RANGE_FILTER_PUSHDOWN

HA_DUPLICATE_KEY_NOT_IN_ORDER

HA_DUPLICATE_POS

HA_ERR_ABORTED_BY_USER

HA_ERR_AUTOINC_ERANGE

HA_ERR_AUTOINC_READ_FAILED

HA_ERR_CANNOT_ADD_FOREIGN

HA_ERR_COMMIT_ERROR

HA_ERR_CORRUPT_EVENT

HA_ERR_CRASHED

HA_ERR_CRASHED_ON_REPAIR

HA_ERR_CRASHED_ON_USAGE

HA_ERR_DECRYPTION_FAILED

HA_ERR_DISK_FULL

HA_ERR_DROP_INDEX_FK

HA_ERR_END_OF_FILE

HA_ERR_ERRORS

HA_ERR_FILE_TOO_SHORT

HA_ERR_FIRST

HA_ERR_FK_DEPTH_EXCEEDED

HA_ERR_FOREIGN_DUPLICATE_KEY

HA_ERR_FOUND_DUPP_KEY

HA_ERR_FOUND_DUPP_UNIQUE

HA_ERR_FTS_TOO_MANY_WORDS_IN_PHRASE

HA_ERR_GENERIC

HA_ERR_INCOMPATIBLE_DEFINITION

HA_ERR_INDEX_COL_TOO_LONG

HA_ERR_INDEX_CORRUPT

HA_ERR_INDEX_FILE_FULL

HA_ERR_INITIALIZATION

HA_ERR_INTERNAL_ERROR

HA_ERR_KEY_NOT_FOUND

HA_ERR_LAST

HA_ERR_LOCK_DEADLOCK

HA_ERR_LOCK_TABLE_FULL

HA_ERR_LOCK_WAIT_TIMEOUT

HA_ERR_LOGGING_IMPOSSIBLE

HA_ERR_NEW_FILE

HA_ERR_NON_UNIQUE_BLOCK_SIZE

HA_ERR_NOT_ALLOWED_COMMAND

HA_ERR_NOT_A_TABLE

HA_ERR_NOT_IN_LOCK_PARTITIONS

HA_ERR_NO_ACTIVE_RECORD

HA_ERR_NO_CONNECTION

HA_ERR_NO_ENCRYPTION

HA_ERR_NO_PARTITION_FOUND

HA_ERR_NO_REFERENCED_ROW

HA_ERR_NO_SAVEPOINT

HA_ERR_NO_SUCH_TABLE

HA_ERR_NULL_IN_SPATIAL

HA_ERR_OLD_FILE

HA_ERR_OUT_OF_MEM

HA_ERR_PARTITION_LIST

HA_ERR_QUERY_INTERRUPTED

HA_ERR_RBR_LOGGING_FAILED

HA_ERR_READ_ONLY_TRANSACTION

HA_ERR_RECORD_CHANGED

HA_ERR_RECORD_DELETED

HA_ERR_RECORD_FILE_FULL

HA_ERR_RECORD_IS_THE_SAME

HA_ERR_RETRY_INIT

HA_ERR_ROWS_EVENT_APPLY

HA_ERR_ROW_IN_WRONG_PARTITION

HA_ERR_ROW_IS_REFERENCED

HA_ERR_ROW_NOT_VISIBLE

HA_ERR_SEQUENCE_INVALID_DATA

HA_ERR_SEQUENCE_RUN_OUT

HA_ERR_TABLESPACE_EXISTS

HA_ERR_TABLESPACE_MISSING

HA_ERR_TABLE_CORRUPT

HA_ERR_TABLE_DEF_CHANGED

HA_ERR_TABLE_EXIST

HA_ERR_TABLE_NEEDS_UPGRADE

HA_ERR_TABLE_READONLY

HA_ERR_TOO_MANY_CONCURRENT_TRXS

HA_ERR_TOO_MANY_FIELDS

HA_ERR_TO_BIG_ROW

HA_ERR_UNDO_REC_TOO_BIG

HA_ERR_UNKNOWN_CHARSET

HA_ERR_UNSUPPORTED

HA_ERR_WRONG_COMMAND

HA_ERR_WRONG_CRC

HA_ERR_WRONG_INDEX

HA_ERR_WRONG_IN_RECORD

HA_ERR_WRONG_MRG_TABLE_DEF

HA_EXT_NOSAME

HA_FAST_CHANGE_PARTITION

HA_FAST_KEY_READ

HA_FILE_BASED

HA_FTS_INVALID_DOCID

HA_FT_MAXBYTELEN

HA_FT_MAXCHARLEN

HA_FT_WLEN

HA_FULLTEXT

HA_GENERATED_KEY

HA_HASH_FIELD_LENGTH

HA_HASH_KEY_LENGTH_WITHOUT_NULL

HA_HASH_KEY_LENGTH_WITH_NULL

HA_HAS_NEW_CHECKSUM

HA_HAS_OLD_CHECKSUM

HA_HAS_OWN_BINLOGGING

HA_HAS_RECORDS

HA_INPLACE_ADD_INDEX_NO_READ_WRITE

HA_INPLACE_ADD_INDEX_NO_WRITE

HA_INPLACE_ADD_PK_INDEX_NO_READ_WRITE

HA_INPLACE_ADD_PK_INDEX_NO_WRITE

HA_INPLACE_ADD_UNIQUE_INDEX_NO_READ_WRITE

HA_INPLACE_ADD_UNIQUE_INDEX_NO_WRITE

HA_INPLACE_DROP_INDEX_NO_READ_WRITE

HA_INPLACE_DROP_INDEX_NO_WRITE

HA_INPLACE_DROP_PK_INDEX_NO_READ_WRITE

HA_INPLACE_DROP_PK_INDEX_NO_WRITE

HA_INPLACE_DROP_UNIQUE_INDEX_NO_READ_WRITE

HA_INPLACE_DROP_UNIQUE_INDEX_NO_WRITE

HA_INVISIBLE_KEY

HA_KEYFLAG_MASK

HA_KEYREAD_ONLY

HA_KEY_BLOB_LENGTH

HA_KEY_HAS_PART_KEY_SEG

HA_KEY_NULL_LENGTH

HA_KEY_SCAN_NOT_ROR

HA_KEY_SWITCH_ALL

HA_KEY_SWITCH_ALL_SAVE

HA_KEY_SWITCH_NONUNIQ

HA_KEY_SWITCH_NONUNIQ_SAVE

HA_LAST_TABLE_FLAG

HA_LEX_CREATE_SEQUENCE

HA_LEX_CREATE_TMP_TABLE

HA_MAX_KEYTYPE

HA_MAX_KEY_BUFF

HA_MAX_KEY_LENGTH

HA_MAX_KEY_SEG

HA_MAX_POSSIBLE_KEY

HA_MAX_POSSIBLE_KEY_BUFF

HA_MAX_REC_LENGTH

HA_MRR_CANT_SORT

HA_MRR_FIXED_KEY

HA_MRR_IMPLEMENTATION_FLAG1

HA_MRR_IMPLEMENTATION_FLAG2

HA_MRR_IMPLEMENTATION_FLAG3

HA_MRR_IMPLEMENTATION_FLAG4

HA_MRR_IMPLEMENTATION_FLAG5

HA_MRR_IMPLEMENTATION_FLAG6

HA_MRR_IMPLEMENTATION_FLAGS

HA_MRR_INDEX_ONLY

HA_MRR_LIMITS

HA_MRR_MATERIALIZED_KEYS

HA_MRR_NO_ASSOCIATION

HA_MRR_NO_NULL_ENDPOINTS

HA_MRR_SINGLE_POINT

HA_MRR_SORTED

HA_MRR_USE_DEFAULT_IMPL

HA_MUST_USE_TABLE_CONDITION_PUSHDOWN

HA_NAMELEN

HA_NEED_READ_RANGE_BUFFER

HA_NON_COMPARABLE_ROWID

HA_NOSAME

HA_NOT_DELETE_WITH_CACHE

HA_NO_AUTO_INCREMENT

HA_NO_BLOBS

HA_NO_COPY_ON_ALTER

HA_NO_ONLINE_ALTER

HA_NO_PREFIX_CHAR_KEYS

HA_NO_SORT

HA_NO_TRANSACTIONS

HA_NULL_ARE_EQUAL

HA_NULL_IN_KEY

HA_NULL_PART

HA_NULL_PART_KEY

HA_ONLINE_ANALYZE

HA_ONLY_WHOLE_INDEX

HA_OPEN_ABORT_IF_CRASHED

HA_OPEN_ABORT_IF_LOCKED

HA_OPEN_COPY

HA_OPEN_DELAY_KEY_WRITE

HA_OPEN_FOR_ALTER

HA_OPEN_FOR_CREATE

HA_OPEN_FOR_DROP

HA_OPEN_FOR_FLUSH

HA_OPEN_FOR_REPAIR

HA_OPEN_FROM_SQL_LAYER

HA_OPEN_IGNORE_IF_LOCKED

HA_OPEN_INTERNAL_TABLE

HA_OPEN_KEYFILE

HA_OPEN_MERGE_TABLE

HA_OPEN_MMAP

HA_OPEN_NO_PSI_CALL

HA_OPEN_TMP_TABLE

HA_OPEN_WAIT_IF_LOCKED

HA_OPTION_CHECKSUM

HA_OPTION_COMPRESS_RECORD

HA_OPTION_DELAY_KEY_WRITE

HA_OPTION_LONG_BLOB_PTR

HA_OPTION_NO_CHECKSUM

HA_OPTION_NO_DELAY_KEY_WRITE

HA_OPTION_NO_PACK_KEYS

HA_OPTION_NO_STATS_PERSISTENT

HA_OPTION_NULL_FIELDS

HA_OPTION_PACK_KEYS

HA_OPTION_PACK_RECORD

HA_OPTION_PAGE_CHECKSUM

HA_OPTION_READ_ONLY_DATA

HA_OPTION_RELIES_ON_SQL_LAYER

HA_OPTION_STATS_PERSISTENT

HA_OPTION_TEMP_COMPRESS_RECORD

HA_OPTION_TEXT_CREATE_OPTIONS_legacy

HA_OPTION_TMP_TABLE

HA_PACK_KEY

HA_PACK_RECORD

HA_PARTIAL_COLUMN_READ

HA_PARTITION_FUNCTION_SUPPORTED

HA_PARTITION_ONE_PHASE

HA_PART_KEY_SEG

HA_PERSISTENT_TABLE

HA_PRESERVE_INSERT_ORDER

HA_PRIMARY_KEY_IN_READ_INDEX

HA_PRIMARY_KEY_REQUIRED_FOR_DELETE

HA_PRIMARY_KEY_REQUIRED_FOR_POSITION

HA_READ_BEFORE_WRITE_REMOVAL

HA_READ_NEXT

HA_READ_ONLY

HA_READ_ORDER

HA_READ_PREV

HA_READ_RANGE

HA_RECORD_MUST_BE_CLEAN_ON_WRITE

HA_REC_NOT_IN_SEQ

HA_REQUIRES_KEY_COLUMNS_FOR_DELETE

HA_REQUIRE_PRIMARY_KEY

HA_REUSES_FILE_NAMES

HA_REVERSE_SORT

HA_SKIP_KEY_SORT

HA_SLOW_CMP_REF

HA_SLOW_RND_POS

HA_SORT_ALLOWS_SAME

HA_SPACE_PACK

HA_SPACE_PACK_USED

HA_SPATIAL

HA_STATE_AKTIV

HA_STATE_BUFF_SAVED

HA_STATE_CHANGED

HA_STATE_DELETED

HA_STATE_EXTEND_BLOCK

HA_STATE_KEY_CHANGED

HA_STATE_NEXT_FOUND

HA_STATE_NO_KEY

HA_STATE_PREV_FOUND

HA_STATE_RNEXT_SAME

HA_STATE_ROW_CHANGED

HA_STATE_WRITE_AT_END

HA_STATE_WRITTEN

HA_STATS_RECORDS_IS_EXACT

HA_STATUS_AUTO

HA_STATUS_CONST

HA_STATUS_ERRKEY

HA_STATUS_NO_LOCK

HA_STATUS_OPEN

HA_STATUS_POS

HA_STATUS_TIME

HA_STATUS_VARIABLE

HA_STATUS_VARIABLE_EXTRA

HA_SWAP_KEY

HA_TABLE_SCAN_ON_INDEX

HA_TRY_READ_ONLY

HA_UNIQUE_HASH

HA_USES_COMMENT

HA_USES_PARSER

HA_VAR_LENGTH_KEY

HA_VAR_LENGTH_PART

HA_VERSIONED_TABLE

HA_WRONG_CREATE_OPTION

INITIAL_LEX_PLUGIN_LIST_SIZE

IO_SIZE

MARIA_PLUGIN_INTERFACE_VERSION

MAX_DATA_LENGTH_FOR_KEY

MAX_REF_PARTS

MYSQL_ANY_PLUGIN

MYSQL_AUDIT_PLUGIN

MYSQL_AUTHENTICATION_PLUGIN

MYSQL_AUTODETECT_CHARSET_NAME

MYSQL_BASE_VERSION

MYSQL_COMPILATION_COMMENT

MYSQL_CONFIG_NAME

MYSQL_DAEMON_INTERFACE_VERSION

MYSQL_DAEMON_PLUGIN

MYSQL_DATADIR

MYSQL_DATA_TRUNCATED

MYSQL_DEFAULT_CHARSET_NAME

MYSQL_DEFAULT_COLLATION_NAME

MYSQL_ERRMSG_SIZE

MYSQL_FTFLAGS_NEED_COPY

MYSQL_FTPARSER_INTERFACE_VERSION

MYSQL_FTPARSER_PLUGIN

MYSQL_HANDLERTON_INTERFACE_VERSION

MYSQL_INFORMATION_SCHEMA_INTERFACE_VERSION

MYSQL_INFORMATION_SCHEMA_PLUGIN

MYSQL_LONG_DATA_HEADER

MYSQL_MAX_PLUGIN_TYPE_NUM

MYSQL_NO_DATA

MYSQL_PLUGIN_INTERFACE_VERSION

MYSQL_PORT

MYSQL_PORT_DEFAULT

MYSQL_REPLICATION_INTERFACE_VERSION

MYSQL_REPLICATION_PLUGIN

MYSQL_SERVER_SUFFIX_DEF

MYSQL_SERVER_VERSION

MYSQL_START_TRANS_OPT_READ_ONLY

MYSQL_START_TRANS_OPT_READ_WRITE

MYSQL_START_TRANS_OPT_WITH_CONS_SNAPSHOT

Options for the START TRANSACTION statement.

MYSQL_STMT_HEADER

MYSQL_STORAGE_ENGINE_PLUGIN

MYSQL_TIME_NOTE_TRUNCATED

MYSQL_TIME_WARN_EDOM

MYSQL_TIME_WARN_NOTES

MYSQL_TIME_WARN_OUT_OF_RANGE

MYSQL_TIME_WARN_TRUNCATED

MYSQL_TIME_WARN_WARNINGS

MYSQL_TIME_WARN_ZERO_DATE

MYSQL_TYPE_VIRTUAL

MYSQL_UDF_PLUGIN

MYSQL_UNIVERSAL_CLIENT_CHARSET

MYSQL_UNIX_ADDR

MYSQL_VALUE_TYPE_INT

MYSQL_VALUE_TYPE_REAL

MYSQL_VALUE_TYPE_STRING

MYSQL_VERSION_EXTRA

MYSQL_VERSION_ID

MYSQL_VERSION_MAJOR

MYSQL_VERSION_MINOR

MYSQL_VERSION_PATCH

MYSQL_WAIT_EXCEPT

MYSQL_WAIT_READ

MYSQL_WAIT_TIMEOUT

MYSQL_WAIT_WRITE

MYSQL_XIDDATASIZE

MYSQL_XID_PREFIX

MYSQL_XID_PREFIX_LEN

MY_AES_BAD_DATA

MY_AES_BAD_KEYSIZE

MY_AES_BLOCK_SIZE

MY_AES_CTX_SIZE

MY_AES_MAX_KEY_LENGTH

MY_AES_OK

MY_AES_OPENSSL_ERROR

MariaDB_DATA_TYPE_PLUGIN

MariaDB_ENCRYPTION_INTERFACE_VERSION

MariaDB_ENCRYPTION_PLUGIN

MariaDB_FUNCTION_PLUGIN

MariaDB_PASSWORD_VALIDATION_PLUGIN

MariaDB_PLUGIN_MATURITY_ALPHA

MariaDB_PLUGIN_MATURITY_BETA

MariaDB_PLUGIN_MATURITY_EXPERIMENTAL

MariaDB_PLUGIN_MATURITY_GAMMA

MariaDB_PLUGIN_MATURITY_STABLE

MariaDB_PLUGIN_MATURITY_UNKNOWN

PLUGINDIR

PLUGIN_INIT_SKIP_INITIALIZATION

PLUGIN_INIT_SKIP_PLUGIN_TABLE

PLUGIN_IS_DELETED

PLUGIN_IS_DISABLED

PLUGIN_IS_DYING

PLUGIN_IS_FREED

PLUGIN_IS_READY

PLUGIN_IS_UNINITIALIZED

PLUGIN_LICENSE_BSD

PLUGIN_LICENSE_BSD_STRING

PLUGIN_LICENSE_GPL

PLUGIN_LICENSE_GPL_STRING

PLUGIN_LICENSE_PROPRIETARY

PLUGIN_LICENSE_PROPRIETARY_STRING

PLUGIN_OPT_NO_INSTALL

PLUGIN_OPT_NO_UNINSTALL

PLUGIN_VAR_BOOL

PLUGIN_VAR_DEPRECATED

PLUGIN_VAR_DOUBLE

PLUGIN_VAR_ENUM

PLUGIN_VAR_INT

PLUGIN_VAR_LONG

PLUGIN_VAR_LONGLONG

PLUGIN_VAR_MASK

PLUGIN_VAR_MEMALLOC

PLUGIN_VAR_NOCMDARG

PLUGIN_VAR_NOCMDOPT

PLUGIN_VAR_NOSYSVAR

PLUGIN_VAR_OPCMDARG

PLUGIN_VAR_READONLY

PLUGIN_VAR_RQCMDARG

PLUGIN_VAR_SET

PLUGIN_VAR_STR

PLUGIN_VAR_THDLOCAL

PLUGIN_VAR_UNSIGNED

my_decimal_test_value

Statics

Bitmap_ALL_BITS_SET

Bitmap_ARRAY_ELEMENTS

MDL_key_m_namespace_to_wait_state_name

MYSQL_PROC_NAME

MYSQL_SCHEMA_NAME

Vers_parse_info_default_end

Vers_parse_info_default_start

Functions

AUTHID_copy⚠

AUTHID_parse⚠

AUTHID_read_from_mysql_proc_row⚠

Alter_inplace_info_Alter_inplace_info⚠

Alter_inplace_info_report_unsupported_error⚠

Used after check_if_supported_inplace_alter() to report error if the result does not match the LOCK/ALGORITHM requirements set by the user.

Binary_string_append⚠

Binary_string_append_longlong⚠

Binary_string_append_ulonglong⚠

Binary_string_copy⚠

Binary_string_copy1⚠

Binary_string_copy2⚠

Binary_string_copy_or_move⚠

Binary_string_copy_printable_hhhh⚠

Convert a string to a printable format. All non-convertable and control characters are replaced to 5-character sequences ‘\hhhh’.

Binary_string_fill⚠

Binary_string_qs_append⚠

Binary_string_qs_append1⚠

Binary_string_qs_append2⚠

Binary_string_qs_append3⚠

Binary_string_qs_append4⚠

Binary_string_qs_append_hex⚠

Binary_string_qs_append_hex_uint32⚠

Binary_string_real_alloc⚠

Binary_string_realloc_raw⚠

Binary_string_replace⚠

Binary_string_reserve⚠

Binary_string_set_fcvt⚠

Binary_string_set_hex⚠

Binary_string_set_hex1⚠

Binary_string_shrink⚠

Binary_string_strrstr⚠

Binary_string_strstr⚠

Binary_string_strstr1⚠

Charset_collation_map_st_from_text⚠

Charset_collation_map_st_insert_or_replace⚠

Charset_collation_map_st_insert_or_replace1⚠

Charset_collation_specific_name⚠

Charset_encoding_allows_reinterpret_as⚠

Charset_eq_collation_specific_names⚠

DTCollation_aggregate⚠

Default_object_creation_ctx_Default_object_creation_ctx⚠

Default_object_creation_ctx_Default_object_creation_ctx1⚠

Default_object_creation_ctx_change_env⚠

Default_object_creation_ctx_create_backup_ctx⚠

File_parser_parse⚠

HA_CREATE_INFO_resolve_to_charset_collation_context⚠

Lex_context_collation_collation_name_for_show⚠

Lex_context_collation_raise_if_not_equal⚠

Lex_exact_charset_extended_collation_attrs_st_merge_collation⚠

Lex_exact_charset_extended_collation_attrs_st_merge_context_collation⚠

Lex_exact_charset_extended_collation_attrs_st_merge_exact_charset⚠

Lex_exact_charset_extended_collation_attrs_st_merge_exact_collation⚠

Lex_exact_charset_extended_collation_attrs_st_resolved_to_character_set⚠

Lex_exact_charset_opt_extended_collate_find_bin_collation⚠

Lex_exact_charset_opt_extended_collate_find_compiled_default_collation⚠

Lex_exact_charset_opt_extended_collate_find_mapped_default_collation⚠

Lex_exact_charset_opt_extended_collate_merge_context_collation⚠

Lex_exact_charset_opt_extended_collate_merge_context_collation_override⚠

Lex_exact_charset_opt_extended_collate_merge_exact_collation⚠

Lex_exact_charset_opt_extended_collate_merge_exact_collation_override⚠

Lex_exact_charset_opt_extended_collate_raise_if_charsets_differ⚠

Lex_exact_charset_opt_extended_collate_raise_if_not_applicable⚠

Lex_exact_charset_raise_if_not_applicable⚠

Lex_exact_charset_raise_if_not_equal⚠

Lex_exact_collation_raise_if_conflicts_with_context_collation⚠

Lex_exact_collation_raise_if_not_equal⚠

Lex_extended_charset_extended_collation_attrs_st_merge_charset_default⚠

Lex_extended_charset_extended_collation_attrs_st_merge_exact_charset⚠

Lex_extended_charset_extended_collation_attrs_st_raise_if_charset_conflicts_with_default⚠

Lex_extended_charset_extended_collation_attrs_st_resolved_to_context⚠

Lex_extended_collation_st_merge⚠

Lex_extended_collation_st_merge_exact_charset⚠

Lex_extended_collation_st_merge_exact_collation⚠

Lex_extended_collation_st_raise_if_conflicts_with_context_collation⚠

Lex_extended_collation_st_set_by_name⚠

Lex_field_type_st_set_handler_length_flags⚠

Lex_length_and_dec_st_set⚠

Log_event_data_type_unpack_optional_attributes⚠

MDL_context_MDL_context⚠

MDL_context_acquire_lock⚠

MDL_context_acquire_locks⚠

MDL_context_clone_ticket⚠

MDL_context_destroy⚠

MDL_context_find_deadlock⚠

MDL_context_get_lock_owner⚠

MDL_context_has_explicit_locks⚠

MDL_context_has_lock⚠

MDL_context_is_lock_owner⚠

MDL_context_release_all_locks_for_name⚠

MDL_context_release_explicit_locks⚠

MDL_context_release_lock⚠

MDL_context_release_statement_locks⚠

MDL_context_release_transactional_locks⚠

MDL_context_rollback_to_savepoint⚠

MDL_context_set_explicit_duration_for_all_locks⚠

MDL_context_set_lock_duration⚠

MDL_context_set_transaction_duration_for_all_locks⚠

MDL_context_try_acquire_lock⚠

MDL_context_upgrade_shared_lock⚠

MDL_context_visit_subgraph⚠

MDL_key_init_psi_keys⚠

MDL_request_init_by_key_with_source⚠

MDL_request_init_with_source⚠

MDL_ticket_accept_visitor⚠

Implement MDL_wait_for_subgraph interface.

MDL_ticket_downgrade_lock⚠

MDL_ticket_get_deadlock_weight⚠

MDL_ticket_get_key⚠

MDL_ticket_get_type_name⚠

MDL_ticket_get_type_name1⚠

MDL_ticket_has_pending_conflicting_lock⚠

MDL_ticket_has_stronger_or_equal_type⚠

MDL_ticket_is_incompatible_when_granted⚠

MDL_ticket_is_incompatible_when_waiting⚠

MDL_ticket_wsrep_report⚠

MDL_wait_MDL_wait⚠

MDL_wait_MDL_wait_destructor⚠

MDL_wait_for_graph_visitor_MDL_wait_for_graph_visitor_destructor⚠

MDL_wait_for_subgraph_MDL_wait_for_subgraph_destructor⚠

MDL_wait_get_status⚠

MDL_wait_reset_status⚠

MDL_wait_set_status⚠

MDL_wait_timed_wait⚠

Natural_join_column_Natural_join_column⚠

Natural_join_column_Natural_join_column1⚠

Natural_join_column_create_item⚠

Natural_join_column_field⚠

Natural_join_column_grant⚠

Natural_join_column_name⚠

Natural_join_column_safe_db_name⚠

Natural_join_column_safe_table_name⚠

Object_creation_ctx_restore_env⚠

Object_creation_ctx_set_n_backup⚠

SEQUENCE_SEQUENCE⚠

SEQUENCE_SEQUENCE_destructor⚠

SEQUENCE_next_value⚠

SEQUENCE_read_initial_values⚠

SEQUENCE_read_lock⚠

SEQUENCE_read_stored_values⚠

SEQUENCE_read_unlock⚠

SEQUENCE_set_value⚠

SEQUENCE_write_lock⚠

SEQUENCE_write_unlock⚠

String_append⚠

String_append1⚠

String_append_for_single_quote⚠

String_append_parenthesized⚠

String_copy⚠

String_copy_aligned⚠

String_needs_conversion⚠

String_needs_conversion_on_storage⚠

String_print⚠

String_print_with_conversion⚠

String_set_ascii⚠

String_set_int⚠

String_set_or_copy_aligned⚠

String_set_real⚠

String_strip_sp⚠

TABLE_LIST_calc_md5⚠

TABLE_LIST_change_refs_to_fields⚠

TABLE_LIST_check_single_table⚠

TABLE_LIST_cleanup_items⚠

TABLE_LIST_containing_subselect⚠

TABLE_LIST_create_field_translation⚠

TABLE_LIST_fetch_number_of_rows⚠

TABLE_LIST_fill_recursive⚠

TABLE_LIST_find_derived_handler⚠

TABLE_LIST_find_underlying_table⚠

TABLE_LIST_find_view_security_context⚠

TABLE_LIST_first_leaf_for_name_resolution⚠

TABLE_LIST_get_first_table⚠

TABLE_LIST_get_real_join_table⚠

TABLE_LIST_get_single_select⚠

TABLE_LIST_get_unit⚠

TABLE_LIST_handle_derived⚠

TABLE_LIST_hide_view_error⚠

TABLE_LIST_init_derived⚠

TABLE_LIST_is_active_sjm⚠

TABLE_LIST_is_leaf_for_name_resolution⚠

TABLE_LIST_is_nonrecursive_derived_with_rec_ref⚠

TABLE_LIST_is_recursive_with_table⚠

TABLE_LIST_is_sjm_scan_table⚠

TABLE_LIST_is_the_same_definition⚠

TABLE_LIST_is_with_table⚠

TABLE_LIST_is_with_table_recursive_reference⚠

TABLE_LIST_last_leaf_for_name_resolution⚠

TABLE_LIST_prepare_security⚠

TABLE_LIST_prepare_view_security_context⚠

TABLE_LIST_print⚠

TABLE_LIST_print_leaf_tables⚠

TABLE_LIST_process_index_hints⚠

TABLE_LIST_register_as_derived_with_rec_ref⚠

TABLE_LIST_register_want_access⚠

TABLE_LIST_reinit_before_use⚠

TABLE_LIST_reset_const_table⚠

TABLE_LIST_set_as_with_table⚠

TABLE_LIST_set_insert_values⚠

TABLE_LIST_set_lock_type⚠

TABLE_LIST_setup_underlying⚠

TABLE_LIST_single_table_updatable⚠

TABLE_LIST_view_check_option⚠

TABLE_LIST_wrap_into_nested_join⚠

TABLE_OPT_RANGE_get_costs⚠

TABLE_OPT_RANGE_index_only_fetch_cost⚠

TABLE_SHARE_actual_n_key_parts⚠

TABLE_SHARE_destroy⚠

Release resources and free memory occupied by the table share.

TABLE_SHARE_free_frm_image⚠

TABLE_SHARE_histograms_exists⚠

TABLE_SHARE_init_from_binary_frm_image⚠

TABLE_SHARE_init_from_sql_statement_string⚠

TABLE_SHARE_init_period_from_extra2⚠

TABLE_SHARE_make_long_hash_func⚠

TABLE_SHARE_read_frm_image⚠

TABLE_SHARE_set_ignored_indexes⚠

TABLE_SHARE_set_overlapped_keys⚠

TABLE_SHARE_update_engine_independent_stats⚠

TABLE_SHARE_update_optimizer_costs⚠

TABLE_SHARE_usable_indexes⚠

TABLE_SHARE_visit_subgraph⚠

TABLE_SHARE_wait_for_old_version⚠

TABLE_SHARE_write_frm_image⚠

TABLE_SHARE_write_par_image⚠

TABLE_STATISTICS_CB_TABLE_STATISTICS_CB⚠

TABLE_STATISTICS_CB_TABLE_STATISTICS_CB_destructor⚠

TABLE_STATISTICS_CB_update_stats_in_table⚠

TABLE_actual_key_flags⚠

TABLE_actual_n_key_parts⚠

TABLE_add_splitting_info_for_key_field⚠

TABLE_add_tmp_key⚠

TABLE_alloc_keys⚠

TABLE_best_range_rowid_filter⚠

TABLE_check_assignability_all_visible_fields⚠

TABLE_check_assignability_explicit_fields⚠

TABLE_check_period_overlaps⚠

TABLE_check_tmp_key⚠

TABLE_check_virtual_columns_marked_for_read⚠

TABLE_check_virtual_columns_marked_for_write⚠

TABLE_clear_column_bitmaps⚠

TABLE_create_key_part_by_field⚠

TABLE_delete_row⚠

TABLE_deny_splitting⚠

TABLE_evaluate_update_default_function⚠

TABLE_export_structure⚠

TABLE_field_to_fill⚠

TABLE_fill_item_list⚠

TABLE_find_constraint_correlated_indexes⚠

TABLE_find_field_by_name⚠

TABLE_free_engine_stats⚠

TABLE_get_materialization_cost⚠

TABLE_init⚠

TABLE_init_cost_info_for_usable_range_rowid_filters⚠

TABLE_init_expr_arena⚠

TABLE_insert_all_rows_into_tmp_table⚠

TABLE_insert_portion_of_time⚠

TABLE_is_filled_at_execution⚠

TABLE_mark_auto_increment_column⚠

TABLE_mark_check_constraint_columns_for_read⚠

TABLE_mark_columns_needed_for_delete⚠

TABLE_mark_columns_needed_for_insert⚠

TABLE_mark_columns_needed_for_update⚠

TABLE_mark_columns_per_binlog_row_image⚠

TABLE_mark_columns_used_by_virtual_fields⚠

TABLE_mark_default_fields_for_write⚠

TABLE_mark_index_columns⚠

TABLE_mark_index_columns_for_read⚠

TABLE_mark_index_columns_no_reset⚠

TABLE_mark_table_for_reopen⚠

TABLE_mark_virtual_columns_for_write⚠

TABLE_move_fields⚠

TABLE_period_make_insert⚠

TABLE_prepare_for_keyread⚠

TABLE_prepare_for_position⚠

TABLE_prepare_triggers_for_delete_stmt_or_event⚠

TABLE_prepare_triggers_for_insert_stmt_or_event⚠

TABLE_prepare_triggers_for_update_stmt_or_event⚠

TABLE_prune_range_rowid_filters⚠

TABLE_remember_blob_values⚠

TABLE_reset_default_fields⚠

TABLE_reset_item_list⚠

TABLE_restore_blob_values⚠

TABLE_restore_column_maps_after_keyread⚠

TABLE_set_spl_opt_info⚠

TABLE_trace_range_rowid_filters⚠

TABLE_update_const_key_parts⚠

TABLE_update_default_fields⚠

TABLE_update_engine_independent_stats⚠

TABLE_update_generated_fields⚠

TABLE_update_virtual_field⚠

TABLE_update_virtual_fields⚠

TABLE_use_index⚠

TABLE_validate_default_values_of_unset_fields⚠

TABLE_vcol_cleanup_expr⚠

TABLE_vcol_fix_expr⚠

TABLE_verify_constraints⚠

TABLE_vers_check_update⚠

TABLE_vers_end_id⚠

TABLE_vers_start_id⚠

TABLE_vers_switch_partition⚠

TABLE_vers_update_end⚠

TABLE_vers_update_fields⚠

Table_period_info_check_field⚠

Table_scope_and_contents_source_st_check_fields⚠

Table_scope_and_contents_source_st_check_period_fields⚠

Table_scope_and_contents_source_st_fix_create_fields⚠

Table_scope_and_contents_source_st_fix_period_fields⚠

Table_scope_and_contents_source_st_vers_check_native⚠

Table_scope_and_contents_source_st_vers_check_system_fields⚠

Table_scope_and_contents_source_st_vers_fix_system_fields⚠

Type_handler_Column_definition_attributes_frm_pack⚠

Type_handler_Column_definition_attributes_frm_unpack⚠

Type_handler_Column_definition_data_type_info_image⚠

Type_handler_Column_definition_prepare_stage1⚠

Type_handler_Column_definition_prepare_stage2_legacy⚠

Type_handler_Column_definition_prepare_stage2_legacy_num⚠

Type_handler_Column_definition_prepare_stage2_legacy_real⚠

Type_handler_Column_definition_redefine_stage1⚠

Type_handler_Column_definition_set_attributes⚠

Type_handler_Column_definition_validate_check_constraint⚠

Type_handler_Item_bool_rowready_func2_fix_length_and_dec⚠

Type_handler_Item_char_typecast_fix_length_and_dec⚠

Type_handler_Item_date_typecast_fix_length_and_dec⚠

Type_handler_Item_datetime_precision⚠

Type_handler_Item_datetime_typecast_fix_length_and_dec⚠

Type_handler_Item_decimal_scale⚠

Type_handler_Item_decimal_typecast_fix_length_and_dec⚠

Type_handler_Item_divisor_precision_increment⚠

Type_handler_Item_double_typecast_fix_length_and_dec⚠

Type_handler_Item_float_typecast_fix_length_and_dec⚠

Type_handler_Item_func_hybrid_field_type_get_date_with_warn⚠

Type_handler_Item_func_min_max_fix_attributes⚠

Type_handler_Item_func_or_sum_illegal_param⚠

Type_handler_Item_func_or_sum_illegal_param1⚠

Type_handler_Item_func_signed_fix_length_and_dec⚠

Type_handler_Item_func_unsigned_fix_length_and_dec⚠

Type_handler_Item_get_date_with_warn⚠

Type_handler_Item_param_set_param_func⚠

Type_handler_Item_param_val_native⚠

Type_handler_Item_send_date⚠

Type_handler_Item_send_datetime⚠

Type_handler_Item_send_double⚠

Type_handler_Item_send_float⚠

Type_handler_Item_send_long⚠

Type_handler_Item_send_longlong⚠

Type_handler_Item_send_short⚠

Type_handler_Item_send_str⚠

Type_handler_Item_send_time⚠

Type_handler_Item_send_timestamp⚠

Type_handler_Item_send_tiny⚠

Type_handler_Item_time_precision⚠

Type_handler_Item_time_typecast_fix_length_and_dec⚠

Type_handler_Key_part_spec_init_foreign⚠

Type_handler_Key_part_spec_init_multiple⚠

Type_handler_Key_part_spec_init_primary⚠

Type_handler_Key_part_spec_init_spatial⚠

Type_handler_Key_part_spec_init_unique⚠

Type_handler_aggregate_for_result_traditional⚠

Type_handler_bit_and_int_mixture_handler⚠

Type_handler_blob_type_handler⚠

Type_handler_blob_type_handler1⚠

Type_handler_calc_key_length⚠

Type_handler_can_return_extract_source⚠

Type_handler_charset_for_protocol⚠

Type_handler_check_null⚠

Type_handler_create_item_copy⚠

Type_handler_get_handler_by_field_type⚠

Type_handler_get_handler_by_real_type⚠

Type_handler_handler_by_log_event_data_type⚠

Type_handler_handler_by_name⚠

Type_handler_handler_by_name_or_error⚠

Type_handler_hybrid_field_type_Type_handler_hybrid_field_type⚠

Type_handler_hybrid_field_type_aggregate_for_comparison⚠

Type_handler_hybrid_field_type_aggregate_for_comparison1⚠

Type_handler_hybrid_field_type_aggregate_for_min_max⚠

Type_handler_hybrid_field_type_aggregate_for_num_op⚠

Type_handler_hybrid_field_type_aggregate_for_result⚠

Type_handler_hybrid_field_type_aggregate_for_result1⚠

Type_handler_is_traditional_scalar_type⚠

Determines MariaDB traditional scalar data types that always present in the server.

Type_handler_make_and_init_table_field⚠

Type_handler_make_num_distinct_aggregator_field⚠

Makes a temporary table Field to handle numeric aggregate functions, e.g. SUM(DISTINCT expr), AVG(DISTINCT expr), etc.

Type_handler_make_packed_sort_key_longlong⚠

Type_handler_make_sort_key_longlong⚠

Type_handler_odbc_literal_type_handler⚠

Type_handler_partition_field_append_value⚠

Type_handler_partition_field_check_result_type⚠

Type_handler_partition_field_type_not_allowed⚠

Type_handler_print_item_value_csstr⚠

Type_handler_print_item_value_temporal⚠

Type_handler_schema⚠

Type_handler_show_binlog_type⚠

Type_handler_store_sort_key_longlong⚠

Type_handler_string_type_handler⚠

Type_handler_type_collection⚠

Type_handler_type_collection_for_aggregation⚠

Type_handler_type_handler_long_or_longlong⚠

Type_handler_validate_implicit_default_value⚠

Type_handler_varstring_type_handler⚠

Return a string type handler for Item If too_big_for_varchar() returns a BLOB variant, according to length. If max_length > 0 create a VARCHAR(n) If max_length == 0 create a CHAR(0)

Type_handler_version⚠

Type_handler_version_mariadb53⚠

Type_handler_version_mysql56⚠

Type_numeric_attributes_aggregate_numeric_attributes_decimal⚠

Type_numeric_attributes_aggregate_numeric_attributes_real⚠

Type_numeric_attributes_count_unsigned⚠

Type_numeric_attributes_find_max_char_length⚠

Type_numeric_attributes_find_max_decimal_int_part⚠

Type_numeric_attributes_find_max_decimals⚠

Type_numeric_attributes_find_max_octet_length⚠

Type_std_attributes_agg_item_collations⚠

Type_std_attributes_agg_item_set_converter⚠

Type_std_attributes_aggregate_attributes_string⚠

Vers_history_point_check_unit⚠

Vers_history_point_eq⚠

Vers_history_point_print⚠

Vers_parse_info_check_conditions⚠

Vers_parse_info_check_sys_fields⚠

Vers_parse_info_create_sys_field⚠

Vers_parse_info_fix_alter_info⚠

Vers_parse_info_fix_create_like⚠

Vers_parse_info_fix_implicit⚠

Vers_parse_info_is_end⚠

Vers_parse_info_is_end1⚠

Vers_parse_info_is_start⚠

Vers_parse_info_is_start1⚠

Vers_parse_info_need_check⚠

View_creation_ctx_create⚠

View_creation_ctx_create1⚠

Wait_for_flush_accept_visitor⚠

Wait_for_flush_get_deadlock_weight⚠

_db_doprnt_⚠

_db_dump_⚠

_db_end_⚠

_db_enter_⚠

_db_explain_⚠

_db_explain_init_⚠

_db_flush_⚠

_db_fp_⚠

_db_get_func_⚠

_db_is_pushed_⚠

_db_keyword_⚠

_db_lock_file_⚠

_db_longjmp_⚠

_db_my_assert⚠

_db_pargs_⚠

_db_pop_⚠

_db_process_⚠

_db_push_⚠

_db_return_⚠

_db_set_⚠

_db_set_init_⚠

_db_setjmp_⚠

_db_suicide_⚠

_db_unlock_file_⚠

base_list_copy⚠

Construct a deep copy of the argument in memory root mem_root. The elements themselves are copied by pointer. If you also need to copy elements by value, you should employ list_copy_and_replace_each_value after creating a copy.

group_by_handler_print_error⚠

ha_bridge_construct⚠

A builder that will create the C++ class from a C vtable. This is used to create a handler from a handlerton.

ha_bridge_destroy⚠

Destroy a handler_bridge

handler_adjust_next_insert_id_after_explicit_value⚠

handler_binlog_log_row⚠

handler_calculate_checksum⚠

handler_change_table_ptr⚠

handler_check_collation_compatibility⚠

handler_check_if_supported_inplace_alter⚠

Check if a storage engine supports a particular alter table in-place

handler_check_if_updates_are_ignored⚠

handler_check_long_hash_compatibility⚠

handler_check_old_types⚠

handler_check_table_binlog_row_based⚠

handler_clone⚠

handler_column_bitmaps_signal⚠

Signal that the table->read_set and table->write_set table maps changed The handler is allowed to set additional bits in the above map in this call. Normally the handler should ignore all calls until we have done a ha_rnd_init() or ha_index_init(), write_row(), update_row or delete_row() as there may be several calls to this routine.

handler_compare_key⚠

handler_compare_key2⚠

handler_compare_key_parts⚠

handler_delete_table⚠

Delete a table in the engine. Called for base as well as temporary tables.

handler_drop_table⚠

handler_end_psi_batch_mode⚠

End a batch started with @c start_psi_batch_mode.

handler_engine_name⚠

handler_get_auto_increment⚠

handler_get_canonical_filename⚠

handler_get_dup_key⚠

handler_get_dynamic_partition_info⚠

handler_get_error_message⚠

handler_get_ha_share_ptr⚠

handler_ha_analyze⚠

handler_ha_bulk_update_row⚠

handler_ha_change_partitions⚠

handler_ha_check⚠

to be actually called to get ‘check()’ functionality

handler_ha_check_and_repair⚠

handler_ha_check_for_upgrade⚠

handler_ha_close⚠

handler_ha_commit_inplace_alter_table⚠

Public function wrapping the actual handler call. Allows us to enforce asserts regardless of handler implementation. @see commit_inplace_alter_table()

handler_ha_create⚠

handler_ha_create_partitioning_metadata⚠

handler_ha_delete_all_rows⚠

handler_ha_delete_row⚠

handler_ha_direct_delete_rows⚠

handler_ha_direct_update_rows⚠

handler_ha_disable_indexes⚠

handler_ha_discard_or_import_tablespace⚠

handler_ha_drop_partitions⚠

handler_ha_drop_table⚠

handler_ha_enable_indexes⚠

handler_ha_end_bulk_insert⚠

handler_ha_external_lock⚠

These functions represent the public interface to users of the handler class, hence they are not virtual. For the inheritance interface, see the (private) functions write_row(), update_row(), and delete_row() below.

handler_ha_index_first⚠

handler_ha_index_last⚠

handler_ha_index_next⚠

handler_ha_index_next_same⚠

handler_ha_index_prev⚠

handler_ha_index_read_idx_map⚠

handler_ha_index_read_map⚠

handler_ha_keyread_clustered_time⚠

handler_ha_keyread_time⚠

handler_ha_open⚠

handler_ha_optimize⚠

handler_ha_prepare_inplace_alter_table⚠

Public functions wrapping the actual handler call. @see prepare_inplace_alter_table()

handler_ha_release_auto_increment⚠

handler_ha_rename_partitions⚠

handler_ha_rename_table⚠

handler_ha_repair⚠

handler_ha_reset⚠

handler_ha_reset_auto_increment⚠

handler_ha_rnd_init_with_error⚠

handler_ha_rnd_next⚠

handler_ha_rnd_pos⚠

handler_ha_table_share_psi⚠

Acquire the instrumented table information from a table share. @return an instrumented table share, or NULL.

handler_ha_thd⚠

handler_ha_truncate⚠

handler_ha_update_row⚠

handler_ha_was_semi_consistent_read⚠

In an UPDATE or DELETE, if the row under the cursor was locked by another transaction, and the engine used an optimistic read of the last committed row value under the cursor, then the engine returns 1 from this function. MySQL must NOT try to update this optimistic value. If the optimistic value does not match the WHERE condition, MySQL can decide to skip over this row. Currently only works for InnoDB. This can be used to avoid unnecessary lock waits.

handler_ha_write_row⚠

handler_index_blocks⚠

handler_index_next_same⚠

handler_index_read_idx_map⚠

Positions an index cursor to the index specified in the handle. Fetches the row if available. If the key value is null, begin at the first key of the index.

handler_is_root_handler⚠

handler_keyread_time⚠

Calculate cost of ‘index_only’ scan for given index, a number of ranges and number of records.

handler_lock_shared_ha_data⚠

handler_log_not_redoable_operation⚠

handler_multi_range_read_info⚠

handler_multi_range_read_info_const⚠

handler_multi_range_read_init⚠

handler_multi_range_read_next⚠

handler_prepare_for_insert⚠

handler_prepare_for_row_logging⚠

handler_print_error⚠

handler_read_first_row⚠

handler_read_range_first⚠

handler_read_range_next⚠

handler_rebind_psi⚠

handler_rename_table⚠

Default rename_table() and delete_table() rename/delete files with a given name and extensions from bas_ext().

handler_set_end_range⚠

handler_set_ha_share_ptr⚠

handler_set_handler_stats⚠

handler_set_lock_type⚠

handler_set_optimizer_costs⚠

handler_start_psi_batch_mode⚠

Put the handler in ‘batch’ mode when collecting table io instrumented events. When operating in batch mode:

handler_unbind_psi⚠

handler_unlock_shared_ha_data⚠

handler_update_auto_increment⚠

handler_update_global_index_stats⚠

handler_update_global_table_stats⚠

handler_use_hidden_primary_key⚠

use_hidden_primary_key() is called in case of an update/delete when (table_flags() and HA_PRIMARY_KEY_REQUIRED_FOR_DELETE) is defined but we don’t have a primary key

my_decimal_my_decimal⚠

my_decimal_to_binary⚠

my_decimal_to_longlong⚠

my_decimal_to_string_native⚠

sequence_definition_adjust_values⚠

sequence_definition_check_and_adjust⚠

sequence_definition_is_allowed_value_type⚠

sequence_definition_prepare_sequence_fields⚠

sequence_definition_read_fields⚠

sequence_definition_store_fields⚠

sequence_definition_truncate_value⚠

sequence_definition_value_type_handler⚠

sequence_definition_value_type_max⚠

sequence_definition_value_type_min⚠

sequence_definition_write⚠

sequence_definition_write_initial_sequence⚠

st_grant_info_all_privilege⚠

st_grant_info_read⚠

st_ha_check_opt_init⚠

st_key_actual_rec_per_key⚠

vers_select_conds_t_check_units⚠

vers_select_conds_t_eq⚠

vers_select_conds_t_init_from_sysvar⚠

vers_select_conds_t_print⚠

Type Aliases

Alter_inplace_info_Rename_keys_vector

Vector of key pairs from DROP/ADD index which can be renamed.

Alter_inplace_info_inplace_alter_table_commit_callback

When ha_commit_inplace_alter_table() is called the the engine can set this to a function to be called after the ddl log is committed.

CHANGE_KEY_CACHE_PARAM

COND

DYNAMIC_ARRAY

Dynamic_array_CMP_FUNC

Dynamic_array_CMP_FUNC2

END_KEY_CACHE

FILE

FLUSH_KEY_BLOCKS

FT_INFO

File

GET_KEY_CACHE_STATISTICS

GRANT_INFO

The current state of the privilege checking process for the current

GRANT_INTERNAL_INFO

State information for internal tables grants. This structure is part of the TABLE_LIST, and is updated during the ACL check process. @sa GRANT_INFO

HANDLER_BUFFER

HASH

HA_CHECK_OPT

INIT_KEY_CACHE

IO_CACHE

IO_CACHE_SHARE

I_P_List_Adapter

I_P_List_Base

Intrusive parameterized list.

I_P_List_Const_Iterator

Iterator for I_P_List.

I_P_List_Iterator

Iterator for I_P_List.

Item_ptr

KEY

KEY_CACHE

KEY_CACHE_FUNCS

KEY_CACHE_INSERT

KEY_CACHE_READ

KEY_CACHE_STATISTICS

KEY_CACHE_WRITE

KEY_MULTI_RANGE

KEY_PART_INFO

LEX_CSTRING

LEX_CUSTRING

LEX_STRING

LIST

Lex_ident

Lex_table_charset_collation_attrs_st

Lex_table_name

List_Iterator_difference_type

List_Iterator_iterator_category

List_Iterator_pointer

List_Iterator_reference

List_Iterator_value_type

List_const_iterator

List_iterator

List_value_type

Log_func

MDL_context_Ticket_iterator

MDL_context_Ticket_list

Intrusive parameterized list.

MDL_request_list

MEM_ROOT

MYSQL

MYSQL_BIND

MYSQL_CONST_LEX_STRING

MYSQL_DATA

MYSQL_FIELD

MYSQL_FIELD_OFFSET

MYSQL_FILE

Type of an instrumented file. @c MYSQL_FILE is a drop-in replacement for @c FILE. @sa mysql_file_open

MYSQL_FTPARSER_BOOLEAN_INFO

MYSQL_FTPARSER_PARAM

MYSQL_LEX_STRING

MYSQL_PARAMETERS

MYSQL_PLUGIN

MYSQL_RES

MYSQL_ROW

MYSQL_ROWS

MYSQL_ROW_OFFSET

MYSQL_SOCKET

An instrumented socket. @c MYSQL_SOCKET is a replacement for @c my_socket.

MYSQL_STMT

MYSQL_THD_KEY_T

MYSQL_TIME

MYSQL_TIME_STATUS

MYSQL_XID

struct st_mysql_xid is binary compatible with the XID structure as in the X/Open CAE Specification, Distributed Transaction Processing: The XA Specification, X/Open Company Ltd., 1991. http://www.opengroup.org/bookstore/catalog/c193.htm

MY_BITMAP

MY_CASEFOLD_CHARACTER

MY_CHARSET_LOADER

MY_CONTRACTION

MY_CONTRACTIONS

MY_DIR

MY_UCA_2BYTES_ITEM

MY_UCA_CONTRACTION_HASH

MY_UCA_INFO

MY_UCA_LEVEL_BOOSTER

MY_UCA_WEIGHT2

MY_UCA_WEIGHT_LEVEL

Mem_root_array_value_type

Convenience typedef, same typedef name as std::vector

NET

ORDER

PSI_memory_key

Instrumented memory key. To instrument memory, a memory key must be obtained using @c register_memory. Using a zero key always disable the instrumentation.

PSI_stage_info

Stage instrument information. @since PSI_VERSION_1 This structure is used to register an instrumented stage.

PSI_stage_key

Instrumented stage key. To instrument a stage, a stage key must be obtained using @c register_stage. Using a zero key always disable the instrumentation.

RANGE_SEQ_IF

REGINFO

RESET_KEY_CACHE_COUNTERS

RESIZE_KEY_CACHE

SELECT_LEX

SELECT_LEX_UNIT

ST_SCHEMA_TABLE

Sql_used_used_t

TABLE_FIELD_DEF

TABLE_FIELD_TYPE

THR_LOCK_DATA

THR_LOCK_INFO

TYPELIB

USED_MEM

Vio

XID

struct xid_t is binary compatible with the XID structure as in the X/Open CAE Specification, Distributed Transaction Processing: The XA Specification, X/Open Company Ltd., 1991. http://www.opengroup.org/bookstore/catalog/c193.htm

_IO_lock_t

__blkcnt_t

__blksize_t

__dev_t

__gid_t

__ino_t

__mode_t

__nlink_t

__off64_t

__off_t

__pthread_list_t

__syscall_slong_t

__time_t

__uid_t

alter_table_operations

decimal_digit_t

decimal_digits_t

decimal_t

intg is the number of decimal digits (NOT number of decimal_digit_t’s !) before the point frac is the number of decimal digits after the point len is the length of buf (length of allocated space) in decimal_digit_t’s, not in bytes sign false means positive, true means negative buf is an array of decimal_digit_t’s

fast_field_copier

field_index_t

ha_checksum

ha_create_table_option

ha_rows

handler_Table_flags

int32

int64

key_map

key_part_map

key_range

lf_pinbox_free_func

longlong

my_bitmap_map

my_bool

my_charset_conv_case

my_charset_conv_mb_wc

my_charset_conv_wc_mb

my_hash_function

my_hash_get_key

my_hash_value_type

my_off_t

my_ptrdiff_t

my_socket

my_thread_id

my_ulonglong

myf

mysql_cond_t

Type of an instrumented condition. @c mysql_cond_t is a drop-in replacement for @c pthread_cond_t. @sa mysql_cond_init @sa mysql_cond_wait @sa mysql_cond_timedwait @sa mysql_cond_signal @sa mysql_cond_broadcast @sa mysql_cond_destroy

mysql_mutex_t

Type of an instrumented mutex. @c mysql_mutex_t is a drop-in replacement for @c pthread_mutex_t. @sa mysql_mutex_assert_owner @sa mysql_mutex_assert_not_owner @sa mysql_mutex_init @sa mysql_mutex_lock @sa mysql_mutex_unlock @sa mysql_mutex_destroy

mysql_prlock_t

Type of an instrumented prlock. A prlock is a read write lock that ‘prefers readers’ (pr). @c mysql_prlock_t is a drop-in replacement for @c rw_pr_lock_t. @sa mysql_prlock_init @sa mysql_prlock_rdlock @sa mysql_prlock_wrlock @sa mysql_prlock_unlock @sa mysql_prlock_destroy

mysql_rwlock_t

Type of an instrumented rwlock. @c mysql_rwlock_t is a drop-in replacement for @c pthread_rwlock_t. @sa mysql_rwlock_init @sa mysql_rwlock_rdlock @sa mysql_rwlock_tryrdlock @sa mysql_rwlock_wrlock @sa mysql_rwlock_trywrlock @sa mysql_rwlock_unlock @sa mysql_rwlock_destroy

mysql_var_check_func

mysql_var_update_func

nested_join_map

page_range

pchar

plugin_ref

pthread_t

qc_engine_callback

query_id_t

range_id_t

range_seq_t

MRR

rw_pr_lock_t

Portable implementation of special type of read-write locks.

stat_print_fn

table_map

time_t

uchar

uint

uint8

uint16

uint32

uint64

ulong

ulonglong