Module mariadb_sys::enum_query_type

Constants

QT_EXPLAIN

This value means focus on readability, not on ability to parse back, etc.

QT_EXPLAIN_EXTENDED

This is used for EXPLAIN EXTENDED extra warnings / Be more detailed Be more detailed than QT_EXPLAIN. Perhaps we should eventually include QT_ITEM_IDENT_SKIP_CURRENT_DATABASE here, as it would give better readable results

QT_FOR_FRM

Print for FRM file. Focus on parse-back. e.g. VIEW expressions and virtual column expressions

QT_ITEM_CACHE_WRAPPER_SKIP_DETAILS

If Item_cache_wrapper should not print <expr_cache>

QT_ITEM_IDENT_DISABLE_DB_TABLE_NAMES

Do not print database name or table name in the identifiers (even if this means the printout will be ambigous). It is assumed that the caller passing this flag knows what they are doing.

QT_ITEM_IDENT_SKIP_DB_NAMES

If identifiers should not include database names, where unambiguous

QT_ITEM_IDENT_SKIP_TABLE_NAMES

If identifiers should not include table names, where unambiguous

QT_ITEM_ORIGINAL_FUNC_NULLIF

If NULLIF(a,b) should print itself as CASE WHEN a_for_comparison=b THEN NULL ELSE a_for_return_value END when “a” was replaced to two different items (e.g. by equal fields propagation in optimize_cond()) or always as NULLIF(a, b). The default behaviour is to use CASE syntax when a_for_return_value is not the same as a_for_comparison. SHOW CREATE {VIEW|PROCEDURE|FUNCTION} and other cases where the original representation is required, should set this flag.

QT_ITEM_SUBSELECT_ID_ONLY

If Item_subselect should print as just “(subquery#1)” rather than display the subquery body

QT_NO_DATA_EXPANSION

good for parsing

QT_NO_WRAPPERS_FOR_TVC_IN_VIEW

This is used for EXPLAIN EXTENDED extra warnings / Be more detailed Be more detailed than QT_EXPLAIN. Perhaps we should eventually include QT_ITEM_IDENT_SKIP_CURRENT_DATABASE here, as it would give better readable results

QT_ORDINARY

Nothing specific, ordinary SQL query.

QT_PARSABLE

good for parsing

QT_SELECT_ONLY

Print for FRM file. Focus on parse-back. e.g. VIEW expressions and virtual column expressions

QT_SHOW_SELECT_NUMBER

This value means focus on readability, not on ability to parse back, etc.

QT_TO_SYSTEM_CHARSET

In utf8.

QT_VIEW_INTERNAL

view internal representation (like QT_ORDINARY except ORDER BY clause)

QT_WITHOUT_INTRODUCERS

Without character set introducers.

Type Aliases

Type

Query type constants (usable as bitmap flags).