Module mariadb_sys::enum_table_category

Constants

TABLE_CATEGORY_INFORMATION

Information schema tables. These tables are an interface provided by the system to inspect the system metadata. These tables do not honor:

TABLE_CATEGORY_LOG

Log tables. These tables are an interface provided by the system to inspect the system logs. These tables do not honor:

TABLE_CATEGORY_PERFORMANCE

Performance schema tables. These tables are an interface provided by the system to inspect the system performance data. These tables do not honor:

TABLE_CATEGORY_SYSTEM

System table, maintained by the server. These tables do honor:

TABLE_CATEGORY_TEMPORARY

Temporary table. The table is visible only in the session. Therefore,

TABLE_CATEGORY_USER

User table. These tables do honor:

TABLE_UNKNOWN_CATEGORY

Unknown value.

Type Aliases

Type

Category of table found in the table share.