Struct mariadb_sys::TABLE_SHARE

#[repr(C)]
pub struct TABLE_SHARE {

    pub table_category: TABLE_CATEGORY,
    pub name_hash: HASH,
    pub mem_root: MEM_ROOT,
    pub keynames: TYPELIB,
    pub fieldnames: TYPELIB,
    pub intervals: *mut TYPELIB,
    pub LOCK_ha_data: mysql_mutex_t,
    pub LOCK_share: mysql_mutex_t,
    pub LOCK_statistics: mysql_mutex_t,
    pub tdc: *mut TDC_element,
    pub tabledef_version: LEX_CUSTRING,
    pub option_list: *mut engine_option_value,
    pub option_struct: *mut ha_table_option_struct,
    pub field: *mut *mut Field,
    pub found_next_number_field: *mut *mut Field,
    pub key_info: *mut KEY,
    pub check_constraints: *mut *mut Virtual_column_info,
    pub blob_field: *mut uint,
    pub vcol_defs: LEX_CUSTRING,
    pub stats_cb: *mut TABLE_STATISTICS_CB,
    pub default_values: *mut uchar,
    pub comment: LEX_CSTRING,
    pub table_charset: *const charset_info_st,
    pub check_set: *mut MY_BITMAP,
    pub all_set: MY_BITMAP,
    pub table_cache_key: LEX_CSTRING,
    pub db: LEX_CSTRING,
    pub table_name: LEX_CSTRING,
    pub path: LEX_CSTRING,
    pub normalized_path: LEX_CSTRING,
    pub connect_string: LEX_CSTRING,
    pub keys_in_use: key_map,
    pub ignored_indexes: key_map,
    pub keys_for_keyread: key_map,
    pub min_rows: ha_rows,
    pub max_rows: ha_rows,
    pub avg_row_length: ulong,
    pub mysql_version: ulong,
    pub reclength: ulong,
    pub stored_rec_length: ulong,
    pub db_plugin: plugin_ref,
    pub optimizer_costs: OPTIMIZER_COSTS,
    pub row_type: Type,
    pub table_type: Type,
    pub tmp_table: Type,
    pub transactional: Type,
    pub page_checksum: Type,
    pub key_block_size: uint,
    pub stats_sample_pages: uint,
    pub stats_auto_recalc: Type,
    pub null_bytes: uint,
    pub last_null_bit_pos: uint,
    pub null_bytes_for_compare: uint,
    pub fields: uint,
    pub stored_fields: uint,
    pub virtual_fields: uint,
    pub virtual_not_stored_blob_fields: uint,
    pub null_fields: uint,
    pub blob_fields: uint,
    pub varchar_fields: uint,
    pub default_fields: uint,
    pub visible_fields: uint,
    pub default_expressions: uint,
    pub table_check_constraints: uint,
    pub field_check_constraints: uint,
    pub rec_buff_length: uint,
    pub keys: uint,
    pub key_parts: uint,
    pub ext_key_parts: uint,
    pub max_key_length: uint,
    pub max_unique_length: uint,
    pub db_create_options: uint,
    pub db_options_in_use: uint,
    pub db_record_offset: uint,
    pub rowid_field_offset: uint,
    pub primary_key: uint,
    pub next_number_index: uint,
    pub next_number_key_offset: uint,
    pub next_number_keypart: uint,
    pub error: Type,
    pub open_errno: uint,
    pub column_bitmap_size: uint,
    pub frm_version: uchar,
    pub check_set_initialized: Type,
    pub use_ext_keys: bool,
    pub null_field_first: bool,
    pub system: bool,
    pub not_usable_by_query_cache: bool,
    pub online_backup: bool,
    pub no_replicate: bool,
    pub crashed: bool,
    pub is_view: bool,
    pub can_cmp_whole_record: bool,
    pub table_creation_was_logged: bool,
    pub non_determinstic_insert: bool,
    pub has_update_default_function: bool,
    pub can_do_row_logging: bool,
    pub long_unique_table: bool,
    pub keep_original_mysql_version: bool,
    pub optimizer_costs_inited: bool,
    pub table_map_id: ulonglong,
    pub incompatible_version: ulong,
    pub view_def: *const File_parser,
    pub sequence: *mut SEQUENCE,
    pub auto_partitioned: bool,
    pub partition_info_str: *mut c_char,
    pub partition_info_str_len: uint,
    pub partition_info_buffer_size: uint,
    pub default_part_plugin: plugin_ref,
    pub online_alter_binlog: *mut Cache_flip_event_log,
    pub versioned: Type,
    pub vers: TABLE_SHARE_period_info_t,
    pub period: TABLE_SHARE_period_info_t,
    pub vers_skip_auto_create: bool,
    pub table_field_def_cache: *const TABLE_FIELD_DEF,
    pub ha_share: *mut Handler_share,
    pub m_psi: *mut PSI_table_share,
    pub frm_image: *mut LEX_CUSTRING,

}

This structure is shared between different table objects. There is one instance of table share per one table in the database.

Fields
table_category: TABLE_CATEGORY

Category of this table.

name_hash: HASHmem_root: MEM_ROOTkeynames: TYPELIBfieldnames: TYPELIBintervals: *mut TYPELIBLOCK_ha_data: mysql_mutex_tLOCK_share: mysql_mutex_tLOCK_statistics: mysql_mutex_ttdc: *mut TDC_elementtabledef_version: LEX_CUSTRINGoption_list: *mut engine_option_valueoption_struct: *mut ha_table_option_structfield: *mut *mut Fieldfound_next_number_field: *mut *mut Fieldkey_info: *mut KEYcheck_constraints: *mut *mut Virtual_column_infoblob_field: *mut uintvcol_defs: LEX_CUSTRINGstats_cb: *mut TABLE_STATISTICS_CBdefault_values: *mut ucharcomment: LEX_CSTRINGtable_charset: *const charset_info_stcheck_set: *mut MY_BITMAPall_set: MY_BITMAPtable_cache_key: LEX_CSTRINGdb: LEX_CSTRINGtable_name: LEX_CSTRINGpath: LEX_CSTRINGnormalized_path: LEX_CSTRINGconnect_string: LEX_CSTRINGkeys_in_use: key_mapignored_indexes: key_mapkeys_for_keyread: key_mapmin_rows: ha_rowsmax_rows: ha_rowsavg_row_length: ulongmysql_version: ulongreclength: ulongstored_rec_length: ulongdb_plugin: plugin_refoptimizer_costs: OPTIMIZER_COSTSrow_type: Typetable_type: Typetmp_table: Typetransactional: Type

Transactional or not.

page_checksum: Type

Per-page checksums or not.

key_block_size: uintstats_sample_pages: uintstats_auto_recalc: Typenull_bytes: uintlast_null_bit_pos: uintnull_bytes_for_compare: uintfields: uintstored_fields: uintvirtual_fields: uintvirtual_not_stored_blob_fields: uintnull_fields: uintblob_fields: uintvarchar_fields: uintdefault_fields: uintvisible_fields: uintdefault_expressions: uinttable_check_constraints: uintfield_check_constraints: uintrec_buff_length: uintkeys: uintkey_parts: uintext_key_parts: uintmax_key_length: uintmax_unique_length: uintdb_create_options: uintdb_options_in_use: uintdb_record_offset: uintrowid_field_offset: uintprimary_key: uintnext_number_index: uintnext_number_key_offset: uintnext_number_keypart: uinterror: Typeopen_errno: uintcolumn_bitmap_size: uintfrm_version: ucharcheck_set_initialized: Typeuse_ext_keys: boolnull_field_first: boolsystem: boolnot_usable_by_query_cache: boolonline_backup: boolno_replicate: boolcrashed: boolis_view: boolcan_cmp_whole_record: booltable_creation_was_logged: boolnon_determinstic_insert: boolhas_update_default_function: boolcan_do_row_logging: boollong_unique_table: boolkeep_original_mysql_version: booloptimizer_costs_inited: booltable_map_id: ulonglongincompatible_version: ulongview_def: *const File_parser

For shares representing views File_parser object with view definition read from .FRM file.

sequence: *mut SEQUENCEauto_partitioned: boolpartition_info_str: *mut c_charpartition_info_str_len: uintpartition_info_buffer_size: uintdefault_part_plugin: plugin_refonline_alter_binlog: *mut Cache_flip_event_logversioned: Typevers: TABLE_SHARE_period_info_tperiod: TABLE_SHARE_period_info_tvers_skip_auto_create: booltable_field_def_cache: *const TABLE_FIELD_DEF

Cache the checked structure of this table.

The pointer data is used to describe the structure that a instance of the table must have. Each element of the array specifies a field that must exist on the table.

The pointer is cached in order to perform the check only once – when the table is loaded from the disk.

ha_share: *mut Handler_share

Main handler’s share

m_psi: *mut PSI_table_share

Instrumentation for this table share.

frm_image: *mut LEX_CUSTRING

only during CREATE TABLE (@sa ha_create_table)

Implementations

impl TABLE_SHARE

pub unsafe fn init_period_from_extra2( &mut self, period: *mut TABLE_SHARE_period_info_t, data: *const uchar, end: *const uchar ) -> bool

pub unsafe fn visit_subgraph( &mut self, waiting_ticket: *mut Wait_for_flush, gvisitor: *mut MDL_wait_for_graph_visitor ) -> bool

pub unsafe fn wait_for_old_version( &mut self, thd: *mut THD, abstime: *mut timespec, deadlock_weight: uint ) -> bool

pub unsafe fn destroy(&mut self)

pub unsafe fn actual_n_key_parts(&mut self, thd: *mut THD) -> uint

pub unsafe fn init_from_binary_frm_image( &mut self, thd: *mut THD, write: bool, frm_image: *const uchar, frm_length: usize, par_image: *const uchar, par_length: usize ) -> c_int

pub unsafe fn init_from_sql_statement_string( &mut self, thd: *mut THD, write: bool, sql: *const c_char, sql_length: usize ) -> c_int

pub unsafe fn write_frm_image( &mut self, frm_image: *const uchar, frm_length: usize ) -> bool

pub unsafe fn write_par_image( &mut self, par_image: *const uchar, par_length: usize ) -> bool

pub unsafe fn read_frm_image( &mut self, frm_image: *mut *const uchar, frm_length: *mut usize ) -> bool

pub unsafe fn free_frm_image(&mut self, frm: *const uchar)

pub unsafe fn set_overlapped_keys(&mut self)

pub unsafe fn set_ignored_indexes(&mut self)

pub unsafe fn usable_indexes(&mut self, thd: *mut THD) -> key_map

pub unsafe fn make_long_hash_func( &self, thd: *mut THD, mem_root: *mut MEM_ROOT, field_list: *mut List ) -> *mut Item_func_hash

pub unsafe fn update_optimizer_costs(&mut self, hton: *mut handlerton)

pub unsafe fn update_engine_independent_stats( &mut self, stat: *mut TABLE_STATISTICS_CB )

pub unsafe fn histograms_exists(&mut self) -> bool