Struct mariadb_sys::Table_scope_and_contents_source_pod_st

#[repr(C)]
pub struct Table_scope_and_contents_source_pod_st {

    pub alter_table_convert_to_charset: *const charset_info_st,
    pub tabledef_version: LEX_CUSTRING,
    pub org_tabledef_version: LEX_CUSTRING,
    pub connect_string: LEX_CSTRING,
    pub comment: LEX_CSTRING,
    pub alias: LEX_CSTRING,
    pub org_storage_engine_name: LEX_CSTRING,
    pub new_storage_engine_name: LEX_CSTRING,
    pub password: *const c_char,
    pub tablespace: *const c_char,
    pub data_file_name: *const c_char,
    pub index_file_name: *const c_char,
    pub max_rows: ulonglong,
    pub min_rows: ulonglong,
    pub auto_increment_value: ulonglong,
    pub table_options: ulong,
    pub avg_row_length: ulong,
    pub used_fields: ulong,
    pub key_block_size: ulong,
    pub expression_length: ulong,
    pub field_check_constraints: ulong,
    pub stats_sample_pages: uint,
    pub null_bits: uint,
    pub options: uint,
    pub merge_insert_method: uint,
    pub extra_size: uint,
    pub db_type: *mut handlerton,
    pub row_type: Type,
    pub transactional: Type,
    pub storage_media: Type,
    pub page_checksum: Type,
    pub option_list: *mut engine_option_value,
    pub stats_auto_recalc: Type,
    pub varchar: bool,
    pub sequence: bool,
    pub check_constraint_list: *mut List,
    pub option_struct: *mut ha_table_option_struct,
    pub fields_option_struct: *mut *mut ha_field_option_struct,
    pub indexes_option_struct: *mut *mut ha_index_option_struct,
    pub table: *mut TABLE,
    pub pos_in_locked_tables: *mut TABLE_LIST,
    pub merge_list: *mut TABLE_LIST,
    pub mdl_ticket: *mut MDL_ticket,
    pub table_was_deleted: bool,
    pub seq_create_info: *mut sequence_definition,

}

A helper struct for table DDL statements, e.g.: CREATE [OR REPLACE] [TEMPORARY] TABLE [IF NOT EXISTS] tbl_name table_contents_source;

Represents a combinations of:

The scope, i.e. TEMPORARY or not TEMPORARY

The “table_contents_source” part of the table DDL statements, which can be initialized from either of these:

table_element_list … // Explicit definition (column and key list)

LIKE another_table_name … // Copy structure from another table

[AS] SELECT … // Copy structure from a subquery

Fields
alter_table_convert_to_charset: *const charset_info_sttabledef_version: LEX_CUSTRINGorg_tabledef_version: LEX_CUSTRINGconnect_string: LEX_CSTRINGcomment: LEX_CSTRINGalias: LEX_CSTRINGorg_storage_engine_name: LEX_CSTRINGnew_storage_engine_name: LEX_CSTRINGpassword: *const c_chartablespace: *const c_chardata_file_name: *const c_charindex_file_name: *const c_charmax_rows: ulonglongmin_rows: ulonglongauto_increment_value: ulonglongtable_options: ulong

HA_OPTION_ values

avg_row_length: ulongused_fields: ulongkey_block_size: ulongexpression_length: ulongfield_check_constraints: ulongstats_sample_pages: uintnull_bits: uintoptions: uintmerge_insert_method: uintextra_size: uintdb_type: *mut handlertonrow_type: Type

Row type of the table definition.

Defaults to ROW_TYPE_DEFAULT for all non-ALTER statements. For ALTER TABLE defaults to ROW_TYPE_NOT_USED (means “keep the current”).

Can be changed either explicitly by the parser. If nothing specified inherits the value of the original table (if present).

transactional: Typestorage_media: Type

DEFAULT, DISK or MEMORY

page_checksum: Type

If we have page_checksums

option_list: *mut engine_option_value

list of table create options

stats_auto_recalc: Typevarchar: bool

1 if table has a VARCHAR

sequence: boolcheck_constraint_list: *mut Listoption_struct: *mut ha_table_option_struct

structure with parsed table options

fields_option_struct: *mut *mut ha_field_option_struct

array of field option structures

indexes_option_struct: *mut *mut ha_index_option_struct

array of index option structures

table: *mut TABLEpos_in_locked_tables: *mut TABLE_LISTmerge_list: *mut TABLE_LISTmdl_ticket: *mut MDL_tickettable_was_deleted: boolseq_create_info: *mut sequence_definition