Type Alias mariadb_sys::GRANT_INFO

pub type GRANT_INFO = st_grant_info;

The current state of the privilege checking process for the current

user, SQL statement and SQL object.

@details The privilege checking process is divided into phases depending on the level of the privilege to be checked and the type of object to be accessed. Due to the mentioned scattering of privilege checking functionality, it is necessary to keep track of the state of the process. This information is stored in privilege, want_privilege, and orig_want_privilege.

A GRANT_INFO also serves as a cache of the privilege hash tables. Relevant members are grant_table and version.

Aliased Type

struct GRANT_INFO {
    pub grant_table_user: *mut GRANT_TABLE,
    pub grant_table_role: *mut GRANT_TABLE,
    pub grant_public: *mut GRANT_TABLE,
    pub version: u32,
    pub privilege: u64,
    pub want_privilege: u64,
    pub orig_want_privilege: u64,
    pub m_internal: st_grant_internal_info,
}

Fields
grant_table_user: *mut GRANT_TABLE

A copy of the privilege information regarding the current host,

database, object and user.

@details The version of this copy is found in GRANT_INFO::version.

grant_table_role: *mut GRANT_TABLEgrant_public: *mut GRANT_TABLEversion: u32

Used for cache invalidation when caching privilege information.

@details The privilege information is stored on disk, with dedicated caches residing in memory: table-level and column-level privileges, respectively, have their own dedicated caches.

The GRANT_INFO works as a level 1 cache with this member updated to the current value of the global variable @c grant_version (@c static variable in sql_acl.cc). It is updated Whenever the GRANT_INFO is refreshed from the level 2 cache. The level 2 cache is the @c column_priv_hash structure (@c static variable in sql_acl.cc)

@see grant_version

privilege: u64

The set of privileges that the current user has fulfilled for a

certain host, database, and object.

@details This field is continually updated throughout the access checking process. In each step the “wanted privilege” is checked against the fulfilled privileges. When/if the intersection of these sets is empty, access is granted.

The set is implemented as a bitmap, with the bits defined in sql_acl.h.

want_privilege: u64

the set of privileges that the current user needs to fulfil in

order to carry out the requested operation.

orig_want_privilege: u64

Stores the requested access acl of top level tables list. Is used to check access rights to the underlying tables of a view.

m_internal: st_grant_internal_info

The grant state for internal tables.